+++ node.gyp
@@ -70,8 +70,10 @@
       'lib/vm.js',
       'lib/zlib.js',
 
+      'lib/internal/child_process.js',
       'lib/internal/freelist.js',
       'lib/internal/smalloc.js',
+      'lib/internal/socket_list.js',
       'lib/internal/repl.js',
     ],
   },
